backend/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
boto3==1.26.84
kafka-python==2.0.2
confluent-kafka==2.1.1
marshmallow==3.19.0
python-dotenv==1.0.0
pydantic==1.10.7
requests==2.28.2
pytest==7.3.1
pytest-flask==1.2.0
pylint==2.17.0
black==23.1.0
gunicorn==20.1.0
flask-restful==0.3.9
prometheus-client==0.16.0
apscheduler==3.10.1
